chore: Move all keys into single file
- Makes accessing the keys easier for invalidations
- Supposedly in v2 of `query-key-factory` we should be able to use more options from `useQuery` such as `staleTime` & `gcTime`
