--- mobile/src/queries/track.ts
@@ -1,20 +1,0 @@
-import { createQueryKeys } from "@lukemorales/query-key-factory";
-
-import { getTrack, getTrackPlaylists, getTracks } from "@/api/new/track";
-
-/** Query keys used in `useQuery` for tracks. */
-export const trackKeys = createQueryKeys("tracks", {
-  all: {
-    queryKey: null,
-    queryFn: () => getTracks(),
-  },
-  detail: (trackId: string) => ({
-    queryKey: [trackId],
-    queryFn: () => getTrack(trackId),
-    contextQueries: {
-      playlists: {
-        queryFn: () => getTrackPlaylists(trackId),
-      },
-    },
-  }),
-});
